+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.1] - 2021-05-15
+### Added
+- Borg commands `serve`, `with-lock`, `break-lock`, and `benchmark crud` as
+  `serve`, `with_lock`, `break_lock`, and `benchmark_crud` respectivly.
+
 ## [0.2.1-alpha.1] - 2021-05-15
 ### Added
 - Added missing options to the following commands: